backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
python-dotenv==1.0.1
python-jose==3.3.0
nltk==3.8.1
textblob==0.17.1
openai==1.12.0
google-generativeai==0.3.2
typing-extensions>=4.8.0
psycopg2==2.9.9 